blostem-pipeline/requirements.txt
fastapi>=0.100
uvicorn>=0.23
sqlmodel>=0.0.13
pydantic>=2.0
langchain>=0.0.300
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
